+++ src/valify/__init__.py
@@ -65,6 +65,6 @@ __all__ = [
     "Schema",    
 ]
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"
 __author__ = "Darshan Bamankar"